small refactoring (with broken database
