+++ src/app/login/page.tsx
@@ -1,4 +1,4 @@
-import LoginForm from "../ui/login-form";
+import LoginForm from "@/components/login-form";
 
 export default function LoginPage() {
   return (
